requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
SQLAlchemy>=2.0
psycopg2-binary>=2.9
alembic>=1.13
pydantic>=2.6
pydantic-settings>=2.2
python-dotenv>=1.0
python-multipart>=0.0.9
httpx[http2]>=0.27
paho-mqtt>=2.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

reportlab>=4.0

# Testing
pytest>=8.0
